move definitions of types to type files. and change reference for these types to reflect the chages
