+++ components/ui/certification-item.tsx
@@ -1,15 +1,6 @@
 import Link from "next/link";
 import React from "react";
 
-type CertificationItemProps = {
-  id: number; // index of certification list
-  name: string; // name of certification
-  organization: string; // organization name
-  date: string; // date of certified
-  verifyLink: string; // link to verify certification
-  className?: string; // additional class name
-};
-
 const CertificationItem = ({
   id,
   name,
@@ -66,4 +57,3 @@ const CertificationItem = ({
 };
 
 export default CertificationItem;
-export type { CertificationItemProps };
