+++ components/ui/nav-bar.tsx
@@ -7,13 +7,6 @@ import { CredlyIcon, GithubIcon, LinkedInIcon, QiitaIcon } from "./icons";
 import ThemeToggle from "../theme-toggle";
 import ChangelogNotification from "./changelog-notification";
 
-interface CustomLinkProps {
-  href: string; // link to the page
-  title: string; // label for the link
-  className?: string; // custom styles for the link
-  prefetch?: boolean; // flag if prefetch is true
-}
-
 const iconSize = "w-6 h-6";
 
 const CustomLink: React.FC<CustomLinkProps> = ({
