+++ types/accordion.tsx
@@ -1,0 +1,7 @@
+type AccordionProps = {
+  index: number;
+  question: string;
+  answer: string;
+  size: string;
+  category?: string;
+};
