+++ types/animated-text.tsx
@@ -1,0 +1,4 @@
+interface AnimatedTextProps {
+  text: string; // text to animate
+  className?: string;
+}
