+++ types/certification-item.tsx
@@ -1,0 +1,8 @@
+type CertificationItemProps = {
+  id: number; // index of certification list
+  name: string; // name of certification
+  organization: string; // organization name
+  date: string; // date of certified
+  verifyLink: string; // link to verify certification
+  className?: string; // additional class name
+};
