+++ types/download-link.tsx
@@ -1,0 +1,7 @@
+type DownLoadLinkProps = {
+  href: string; // link to the page
+  target?: string; // target for the link Optional
+  className?: string; // custom styles for the link Optional
+  download?: boolean; // flag if download is true Optional
+  label: string; // label for the link
+};
