+++ types/education-history.tsx
@@ -1,0 +1,7 @@
+type EducationHistoryProps = {
+  school: string;
+  department: string;
+  startYear: string;
+  endYear: string;
+  description: string;
+};
