+++ types/faq.tsx
@@ -1,0 +1,7 @@
+type FaqProps = {
+  // id: number; exclude id for maintainability
+  question: string;
+  answer: string;
+  size: string;
+  category: string;
+};
