+++ types/nav-bar.tsx
@@ -1,0 +1,6 @@
+interface CustomLinkProps {
+  href: string; // link to the page
+  title: string; // label for the link
+  className?: string; // custom styles for the link
+  prefetch?: boolean; // flag if prefetch is true
+}
