+++ types/project-item.tsx
@@ -1,0 +1,9 @@
+type Project = {
+  // id: string; exclude id for maintainability
+  title: string;
+  description: string;
+  image: string;
+  technologies: string[];
+  githubUrl: string;
+  liveUrl: string;
+};
