+++ types/skill-category.tsx
@@ -1,0 +1,5 @@
+type SkillCategoryProps = {
+  category: string;
+  selectedCategories: Set<string>;
+  onClick: (category: string) => void;
+};
