+++ types/skill-item.tsx
@@ -1,0 +1,6 @@
+type SkillItemProps = {
+  name: string;
+  category: string;
+  years: string;
+  proficyency?: string; // onBusiness or self-study
+};
